+++ pkg/providers/property/ruleformats/rule_format_v2024_01_09.gen.go
@@ -9110,7 +9110,7 @@ func getBehaviorsSchemaV20240109() map[string]*schema.Schema {
 		"origin_characteristics": {
 			Optional:    true,
 			Type:        schema.TypeList,
-			Description: "Specifies characteristics of the origin. Akamai uses this information to optimize your metadata configuration, which may result in better origin offload and end-user performance. This behavior cannot be used in includes.",
+			Description: "Specifies characteristics of the origin. Akamai uses this information to optimize your metadata configuration, which may result in better origin offload and end-user performance. This behavior can be used in includes.",
 			MaxItems:    1,
 			Elem: &schema.Resource{
 				Schema: map[string]*schema.Schema{
@@ -10280,7 +10280,7 @@ func getBehaviorsSchemaV20240109() map[string]*schema.Schema {
 		"read_timeout": {
 			Optional:    true,
 			Type:        schema.TypeList,
-			Description: "This behavior specifies how long the edge server should wait for a response from the requesting forward server after a connection has already been established. Any failure to read aborts the request and sends a `504` Gateway Timeout error to the client. Contact Akamai Professional Services for help configuring this behavior. This behavior can be used in includes.",
+			Description: "This behavior specifies how long the edge server should wait for a response from the requesting forward server after a connection has already been established. This behavior can be used in includes.",
 			MaxItems:    1,
 			Elem: &schema.Resource{
 				Schema: map[string]*schema.Schema{
@@ -10303,7 +10303,7 @@ func getBehaviorsSchemaV20240109() map[string]*schema.Schema {
 					"value": {
 						ValidateDiagFunc: validateRegexOrVariable("^[0-9]+[DdHhMmSs]$"),
 						Optional:         true,
-						Description:      "Specifies the read timeout necessary before failing with a `504` error. This value should never be zero.",
+						Description:      "The amount of time an edge server should wait for each read statement to return a response from the forward server after a connection has already been established. Larger objects may need many reads, and this timeout applies to each read separately. Any failure to complete a read within this time limit aborts the request and sends a 504 Gateway Timeout error to the client.",
 						Type:             schema.TypeString,
 					},
 				},
@@ -11677,7 +11677,7 @@ func getBehaviorsSchemaV20240109() map[string]*schema.Schema {
 		"script_management": {
 			Optional:    true,
 			Type:        schema.TypeList,
-			Description: "Ensures unresponsive linked JavaScript files do not prevent HTML pages from loading. This behavior cannot be used in includes.",
+			Description: "Ensures unresponsive linked JavaScript files do not prevent HTML pages from loading. See `Script Management API` for more information. This behavior cannot be used in includes.",
 			MaxItems:    1,
 			Elem: &schema.Resource{
 				Schema: map[string]*schema.Schema{